backend/requirements.txt
alembic==1.11.3
annotated-types==0.5.0
anyio==3.7.1
arrow==1.2.3
binaryornot==0.4.4
certifi==2023.5.7
chardet==5.1.0
charset-normalizer==3.2.0
click==8.1.4
cookiecutter==1.7.3
dnspython==2.3.0
ecdsa==0.18.0
email-validator==2.0.0.post2
exceptiongroup==1.1.2
fastapi==0.100.0
fastapi-utils==0.2.1
h11==0.14.0
idna==3.4
install==1.3.5
Jinja2==3.1.2
jinja2-time==0.2.0
Mako==1.2.4
manage-fastapi==1.1.1
MarkupSafe==2.1.3
poyo==0.5.0
prompt-toolkit==3.0.39
psycopg2-binary==2.9.7
pyasn1==0.5.0
pydantic==1.10.11
pydantic_core==2.1.2
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
python-slugify==8.0.1
questionary==1.10.0
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.41
sqlalchemy2-stubs==0.0.2a35
sqlmodel==0.0.8
starlette==0.27.0
text-unidecode==1.3
typer==0.4.2
typing_extensions==4.7.1
urllib3==2.0.3
uvicorn==0.23.2
wcwidth==0.2.6
websockets==11.0.3
